+++ agvw-like/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.0.1
+
+- Misc
+  - `GameContent` の `argument` を `any` に修正
+
 ## 1.0.0
 
 公開